Fixed many warnings.

+++ source/Core/Castor3D/Model/Mesh/Importer.cpp
@@ -142,9 +142,9 @@ namespace castor3d
 			float br = getHeight( data, width, height, x + 1, y + 1 ); // bottom right
 
 			// sobel filter
-			const float dX = ( tr + 2.0 * r + br ) - ( tl + 2.0 * l + bl );
-			const float dY = ( bl + 2.0 * b + br ) - ( tl + 2.0 * t + tr );
-			const float dZ = 1.0 / normalStrength;
+			const float dX = float( ( tr + 2.0 * r + br ) - ( tl + 2.0 * l + bl ) );
+			const float dY = float( ( bl + 2.0 * b + br ) - ( tl + 2.0 * t + tr ) );
+			const float dZ = float( 1.0 / normalStrength );
 
 			castor::Point3f n{ dX, dY, dZ };
 			castor::point::normalise( n );
